feat: Move stardust NodeInfoService to stardust folder
